kernel: re-organize process.rs
