+++ kernel/src/memop.rs
@@ -1,6 +1,6 @@
 //! Implementation of the MEMOP family of syscalls.
 
-use crate::process::ProcessType;
+use crate::process::Process;
 use crate::syscall::SyscallReturn;
 use crate::ErrorCode;
 
@@ -37,7 +37,7 @@ use crate::ErrorCode;
 ///   where the app has put the start of its heap. This is not strictly
 ///   necessary for correct operation, but allows for better debugging if the
 ///   app crashes.
-pub(crate) fn memop(process: &dyn ProcessType, op_type: usize, r1: usize) -> SyscallReturn {
+pub(crate) fn memop(process: &dyn Process, op_type: usize, r1: usize) -> SyscallReturn {
     match op_type {
         // Op Type 0: BRK
         0 /* BRK */ => {
